Facebook changes to xtrabackup
Summary:
We build this in our tree, beside our normal out-of-dir cmake build.

Basically, this can be built just by running this on a clean source tree:

	cd xtrabackup; ./utils/build.sh innodb56; cd ..

Note that this only works right in a clean source tree.
But, since we build out-of-dir, our source tree is always clean.

Also note that some corner-cases (Debug builds of xtrabackup, for example)
will probably still require more build magic to work.  The real solution
here would be to integrate this build into the main cmake build of
mysql-5.6 itself, but we didn't go that far with this (yet).

Sections not required for mysql-5.6 were removed, for simplicity.

Also, some compile/build warnings/errors were cleaned up.

And, innobackupex was modified.  For apply log, if xtrabackup binary is
explicitly specified, dont try to autodetect.

Also, changes for 5.6.10 -> 5.6.11:
 * The os_file_delete() macro/function got a new argument, added that.
 * The call to fil_mtr_rename_log was moved, moved the #ifdef also.

Test Plan:
xtrabackup build works (in CLEAN source dir):
	cd xtrabackup; ./utils/build.sh innodb56; cd ..

Reviewers: rudradevbasak, jtolmer

Reviewed By: rudradevbasak, jtolmer