Port testing of random compression failures
Summary:
This is a port of D382101 and D384911 from 5.1. Changes made from
the original revisions are that the I took the code in page0zip.cc
from HEAD of origin/prod2012q3 and I had to change the check to
account for 5.6's built in padding vs. our padding in 5.1. Also,
I had to change ocurrences of innodb_file_format_check to
innodb_file_format_max in the tests.

Test Plan: arc unit / Jenkins

Reviewers: nizamordulu, rahulgulati

Reviewed By: nizamordulu

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D756794

Task ID: 2002422