Share more structures
Summary:
This is a base change required for lots of stuff,
including innodb_fake_changes, stats, etc....

I am factoring this out to allow diff re-ordering.

Test Plan:
Build is clean, and so are mtr tests.

This code was factored out of the existing basic stats diff.
There are no overall code changes once that diff is applied.

Reviewers: jtolmer, chip

Reviewed By: chip

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D757490

Task ID: 2236044