Make the xtrabackup build hack more universal
Summary:
This allows it to build without any of our build stuff.
It also doesn't break our method of building.

This diff will be squashed with the man xtrabackup diff.

Test Plan:
Built it with our stuff and without our stuff.
Built it on my devserver and on my laptop.
Built it out-of-dir and in-dir.

Reviewers: rudradevbasak

Reviewed By: rudradevbasak

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D757601

Task ID: 2151092