Port v5.1 innodb stress tests to 5.6
Summary:
This is a port of all stress, bigstress, and hugestress tests to 5.6.

This is a complete port, except:

Removed all *_gc.test.  These tested a group commit mechanism that
is not in 5.6, and will never be ported to it.

This can stress test secondary indexes as well, but this is disabled.

Removed extra innodb_file_format_check.   It is no longer needed in 5.6.

Updated results files (partly with --record, partly manually).

Removed use of vars we've not (yet) ported.  These included:
	innodb_zlib_wrap
	innodb_log_compressed_pages
	innodb_background_checkpoint
	innodb_prepare_commit_mutex
	innodb_prefix_index_cluster_optimization

Updated and verified changes to all results files.

Test Plan:
Ran all the non-huge tests, and one huge test, myself, so far so good.
Jenkins nightlies will start hammering these after this is pushed.

Reviewers: nizamordulu, chip, jtolmer

Reviewed By: jtolmer