Backport Oracle's fix for the buddy allocator free list scan.
Summary:
Oracle's patch for to fix the buddy allocator performance
problem of needing to do a linear scan of zip_free, which
caused us problems when trying to use a key block size of
4k.

Test Plan:
mtr, dump and parallel reload of production instances.

Internal:
Oracle's patch file, acquired from the SR, can be found at
/home/jtolmer/patches/bug16223169-v7.patch.

Their fix is supposedly already pushed to 5.6.11 so this diff
should not need to be ported forward.

Reviewers: nizamordulu, steaphan

Reviewed By: steaphan

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D769954

Task ID: 2247891