Port v5.1 Cache mem_block_t Memory Allocations
Summary:
We modify mem_block_struct to support caching of the mem_block_t
objects. Note that mem_heap_t is a linked list of allocated memory
blocks.

A call to mem_heap_create_cached() produces a cached heap only if the
requested initial memory size doesn't exceed the cache block size. If
that is not the case than none of the blocks in the returned heap is
cached. This makes it important to have good cache block size estimates.

If the initial block in a heap is cached (i.e. mem_heap_create_cached
returns a heap whose heap->cache is not NULL) then consecutive blocks
are cached if the consecutive requested memory sizes do not exceed the
cache block size. (see the change in the implementation of
mem_heap_add_block() for details) It's possible for a heap to have some
of its blocks cached and some of them not cached. Non-cached blocks will
have block->cache=NULL.

The rule of thumb is that if block->cache is not NULL then the block
must be put back to cache when the heap is freed unless the cache is
full. If the cache is full then the block gets free()'d.

This diff also fixes the memory usage estimations used in
page_zip_compress() page_zip_decompress(). Earlier estimations did not
take into account the size of deflate_state and the size of
inflate_state respectively. I went through the ZALLOC() calls in
deflate.c and inflate.c to obtain the new estimations.

Also fix the tests innodb_log_compressed_pages* and
innodb_change_compression_level* because their result file depended on
the default value of innodb_log_compressed_pages.

We use the same malloc_cache object to cache both page compressions and
blob compressions. blob compressions allocate 250K whereas page
compressions require about 350K.
We also use the same malloc_cache for blob decompressions and page
decompressions. blob decompressions allocate 40K and page decompressions
allocate around 50K.

Change memLevel for blob compressions from 7 to 8. I made this change
because:
1- it's backwards compatible. we don't need to recompress the data in
the existing tables.
2- Since we are using malloc_cache we already have excess memory in
cached malloc blocks. By using a higher memLevel we can get better
compression for blobs at no cost.

Internal:
innodb_cache_malloc populates a table and looks at the hits and
misses for the memory allocation caches that's used for page compression
and decompression.

Assuming the cache size is large enough, the number of misses must not
be greater than the size of the cache. Eg. if there is only one thread
doing compressions then a cache size of 1 is enough because it will
acquire a cached memory block, use it and put it back after it's done.
Next time it needs a block, there will be one in the cache.

In 5.6.12 there are two threads doing decompressions as opposed to one:
the user thread and the dict_stats_thread. For this reason the cache for
decompression blocks must be at least 2.

This is a port of the following diff: https://phabricator.fb.com/D418406

Test Plan: mysqltest.sh, arc unit (Jenkins)

Reviewers: nizamordulu

Reviewed By: nizamordulu

CC: mcallaghan, jtolmer, db-eng@lists

Differential Revision: https://phabricator.fb.com/D765901
Differential Revision: https://phabricator.fb.com/D823746

Task ID: 2002402

Conflicts:
	storage/innobase/handler/ha_innodb.cc