Reduce LOCK_log mutex contention during lock during binlog reads.
Summary:
This removes acquisition of LOCK_log mutex during binlog reads.

binlog_last_valid_pos is set whenever binlog event is written to
log_file.  When binlog dump threads read from the binlog file, EOF is
changed to the previous stable position by using
get_binlog_last_valid_pos() which causes it to read complete or valid
events. If there are no further complete events to read, it waits for
signal_update() before starting reading from the binlog again. Also if
reading from inactive binlogs(binlogs which already got rotated and no
events will be written) mutex is never held.

This also happens to fix: http://bugs.mysql.com/61545

This changes the bin log dump thread so that it does not need to
acquire LOCK_log.  Because of that, it was possible for the dump
thread to read events from the bin log before the after_flush hook
is called.  When the dump thread wins that race, it means that
ReplSemiSyncMaster::writeTranxInBinlog was not yet called and so
the dump thread will not request an ACK from the slave for the
group being committed, possibly resulting in the master waiting
for the full semi-sync timeout to expire before continuing.
So, this also fixes that race.

Test Plan: mtr.

Reviewers: jtolmer

Reviewed By: jtolmer