Disable binlog.binlog_server_start_options
Summary:
The test has a few flaws. First, it does not support
running multiple instances concurrently because it writes
and removes a cnf file in a shared location. Second, the datadir
it creates assumes that the server was built without specifying
a default location for datadir other than 'data/'. Also, because
it uses the compiled-in datadir, that also prevents it from being
able to run multiple instances of the test in parallel. Finally,
if one does try to run multiple instances of the test in parallel,
the test may end up doing a rmtree('/data') which may be bad
if '/data' exists.

Test Plan: None

Reviewers: steaphan

Reviewed By: steaphan

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D782235

Conflicts:
	mysql-test/suite/binlog/t/disabled.def