Two fixes for logical read-ahead feature
Summary:
1) Do not use adaptive hash index for node pointer records.
2) Fix logical read-ahead for debug build.

This diff is meant to fix two bugs related to logical read ahead:

1- Compute the depth of B-tree by looking at the height of the root
node. Earlier we were searching for a record and reading the value from
btr_cursor.tree_height but this field is not set when adaptive hash
index is used.

2- When UNIV_DEBUG is enabled trx->lra_sort_arr array was zeroed before
it was filled with new page numbers. A previous revision changed the
sort array size but I forgot to change the size parameter used for
memset().

Test Plan:
* mtr
Reviewers: yoshinori, steaphan

Reviewed By: yoshinori