Fix disable_deadlock_detect
Summary:
The initial port of innodb_deadlock_detect was completely broken.
Rather than disable deadlock detection, it just ignored the results of it.

This flaw was found and reported by: github.com/zhaiwx1987

It is unclear if this is still a useful feature, but this fixes it.

Internal:
This should have been part of:
	https://phabricator.fb.com/D462795

Test Plan: Jenkins

Reviewers: jtolmer

Reviewed By: jtolmer

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D794014