Create xtrabackup_logfile in the supplied target_dir
Summary:
Without this change, xtrabackup_logfile is created in the directory specified as tmpdir in my.cnf file.
As the logfile can be quite large (>10 GB) the tmpdir is not always the best option. Hence, create xtrabackup_logfile in the supplied target_dir.

Internal:
With the change in D810030, this will now create xtrabackup_logfile in /data/xtra3307 directory

Test Plan: Checked with lsof | grep xtrabackup_logfile, that the logfile is created in the directory specified with --tmpdir in innobackupex and --target_dir in xtrabackup

Reviewers: steaphan

Reviewed By: steaphan

CC: db-eng@lists, mysql-eng@

Differential Revision: https://phabricator.fb.com/D810029

Task ID: 2382338
(cherry picked from commit 2a215acd17632372c541955fed7d2cad2694fa31)