Create xtrabackup_logfile in the supplied target_dir
Summary:
Without this change, xtrabackup_logfile is created in the directory
specified as tmpdir in my.cnf file.

As the logfile can be quite large (>10 GB) the tmpdir is not always the
best option. Hence, create xtrabackup_logfile in the supplied
target_dir.

Test Plan:
Checked with lsof | grep xtrabackup_logfile, that the logfile is created
in the directory specified with --tmpdir in innobackupex and
--target_dir in xtrabackup

Reviewers: steaphan

Reviewed By: steaphan