Port v5.1 "Improve page_zip_validate()" to 5.6
Summary:
InnoDB compresses the unused parts of the uncompressed page (garbage).
It does this because page_zip_validate() compares the uncompressed page
data against the compressed page data byte by byte. We improve
page_zip_validate by comparing only the used record data as
opposed to the raw data.

When UNIV_DEBUG is enabled, this revision adds code to clean the
garbage parts of the uncompressed page before it is compressed. This
makes sure that the new page_zip_validate() works as expected because
the garbage parts of the original uncompressed page and the garbage
parts of the uncompressed version of the newly compressed page do not
match.

Note that the page format doesn't change except when UNIV_DEBUG is
enabled and it is not enabled in production so we are good.

Test Plan:
Jenkins (arc unit), all tests pass.
Build with UNIV_ZIP_DEBUG enabled and run innodb_stress.

Reviewers: nizamordulu

Reviewed By: nizamordulu