Port v5.1 "Improve page_zip_validate()" to 5.6
Summary:
InnoDB compresses the unused parts of the uncompressed page (garbage).
It does this because page_zip_validate() compares the uncompressed page
data against the compressed page data byte by byte. We improve
page_zip_validate by comparing only the used record data as
opposed to the raw data.

When UNIV_DEBUG is enabled, this revision adds code to clean the
garbage parts of the uncompressed page before it is compressed. This
makes sure that the new page_zip_validate() works as expected because
the garbage parts of the original uncompressed page and the garbage
parts of the uncompressed version of the newly compressed page do not
match.

Note that the page format doesn't change except when UNIV_DEBUG is
enabled and it is not enabled in production so we are good.

Internal:
This is a port of diff D598172
(commit# rMSFOT7582b228743022b78965d79169c3081e07c3adca)

This is a direct port except,

(1) added an #ifndef UNIV_INNOCHECKSUM for heap_ptr and index_ptr
(defined in page0zip.h), which was causing build failures for
Innochecksum.

(2) use static_cast to convert void* to ib_page_t* from call to
mem_heap_zalloc() at page0zip.cc:1527.

(3) minor indentation changes.

Test Plan: Jenkins (arc unit)

Reviewers: nizamordulu

Reviewed By: nizamordulu

CC: jtolmer, mysql-eng@

Differential Revision: https://phabricator.fb.com/D816443

Task ID: 2270168