Disable extra checksum check on startup (temporary).
Summary:
Something is wrong with this.  Until I figure out what and fix it,
I am disabling this check.  It was never done here before, so we are
not losing anything we had by disabling it for now.

In order to keep the diff minimal, I just assigned to NULL (the return
value from this new function in case of success), instead of the call.

To get rid of the "defined but not used" compiler warning for this,
I actually removed the entire function, as this was the only call to it.

Also had to revert the test change from 5.6.12 that looks for this
new error message, instead of the old one from 5.6.11.

This will be reverted and replaced with the real fix when it's done.

@internal

Test Plan: Yoshi said it worked.

Reviewers: jtolmer

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D821577

Task ID: 2420628