Refuse gtid_mode=ON when replication rules on db or tables are turned on
Summary:
Refuse gtid_mode=ON when replication rules on db or tables are
turned on.

Test Plan:
Tested by running mysqld with this change with gtid_mode=ON
and turning on replication rules. mysqld failed to start every time with
expected error. mysqld restarted successfully without any replication
rules. Modified a test which is using replicate-ignore-table along with
gtid_mode=ON.

Reviewers: wultsch, jtolmer

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D818645

Task ID: 2371011