With MTS sql_thread crashes during slave binlog rotate
Summary:
With multi threaded slave turned on sql_thread crashes consistently
during slave binlog rotate. This is due to one of the slave worker
threads failing during rotate which then signals sql_thread to die (This
is by design of MTS where if any slave worker gets an error during
commit, the coordinator will die). The failure during rotate is indeed
not a failure but a bug in error handling changes from 5.6.11 to 5.6.12.

Note that the bug doesn't cause sql_thread to die in single threaded
replication mode which is why this is only reproducible with MTS.

The public bug report is bugs.mysql.com/bug.php?id=69369

Test Plan:
Added an mtr test case with MTS turned on and with
max_binlog_size set appropriately so that a rotate of slave binlog
happens during the test. Added GTID cnf options also to test both
GTID and MTS simultaneously, since there are no such tests in test
suite (but the bug is nothing to do with GTID feauture).

Reviewers: jtolmer, steaphan

Reviewed By: jtolmer