Add fix for deleted tablespaces in xtrabackup apply log
Summary:
Without this fix xtrabackup apply log may hang at the end, if it failed
to open some .ibd file in the past.

This patch was part of the old innodb51.patch supplied by Percona, but
for some weird reason it is not part of their current innodb56.patch on
which we have based our patches.

Test Plan:
Added test where table is modified and then dropped during backup.
Succeeds with this patch, fails otherwise.

Reviewers: steaphan

Reviewed By: steaphan