Added count of bytes of comments to table_statistics
Summary:
Added a new field "COMMENT_BYTES" to the information_schema.table_statistics table which counts the bytes of 4 types of comments in queries , per table.The 4 types are :
	1. C style comments, most commonly used : /*   */
	2. Comments beginning with "#" to the end of line
	3. Comments beginning wiht "-- " to the end of line
    4. Versionedn comments : /*!xxxxx   */
                                 In this type of comment , xxxxx is supposed to represent the version of MySQL server currently being used.If the current version is greater than or equal to the
                                 xxxxx value , then this is not treated as a comment and the syntax within it is executed . However if the current running version of MySQL
                                 is less than this xxxxx value , then this whole thing is treated as a comment.

	Reference : http://dev.mysql.com/doc/refman/5.6/en/comments.html

For a join query , the comment bytes are added to the statistics of all the tables involved in the join.
This also supports multiple hyrbid comments in a single query.

Test Plan:
count_bytes_comments.test
	   jenkins

Reviewers: rongrong, steaphan

Reviewed By: rongrong