crash safe slave when using GTID and MTS
Summary:
Gtid information is stored in a transaction table
(slave_gtid_info). The table contains three columns

  Id  | Database name  | Last_gtid

Every database contains a row in the table storing the last gtid
in the replication stream for that database. On every commit by the
sql_thread or slave worker thread in case of MTS, the correponding
row in the table slave_gtid_info is updated using XA. The table schema
slave_worker_info is changed to include last_gtid executed by that worker.
This extra information helps in finding out the group where worker stopped
in case of an error.

Every group gtid is checked with the last_gtid seen for that database to
decide whether it should be executed or skipped in which case only a empty
group of that gtid is printed in the binary log. Empty group in the binlog looks like
this

  SET @@session.GTID_NEXT='';
  BEGIN;
  // Followup of this diff (either wormhole comments or the whole event)
  COMMIT;

The follow up is to add wormhole comments inside the empty group or even better
the whole event without actually committing in Innodb. The empty groups fills the
gaps in gtid_executed set on slave. This helps avoiding the situations where master
needs to scan old binlogs to resend the gaps again.

Current setting for crash safe slave with gtid and mts requires

  --relay_log_info_repository='TABLE'
  (currently this won't work with MTS, a restart with relay-log-recovery=0 followed by
  START SLAVE UNTIL SQL_AFTER_MTS_GAPS is required)
  --relay_log_recovery=1
  --gtid_mode=ON

without --relay_log_recovery=1 a transaction can get lost if the relay log ends with GTID
event. With this transactional table the new crash safe slave parameters with gtid and
MTS requires.

  --relay_log_recovery=1
  --gtid_mode=ON

This avoids the need for START SLAVE UNTIL SQL_AFTER_MTS_GAPS command and sync_relay_log=1.
The disadvantage is to copy binlogs again from master if slave is severely lagged.

Test Plan:
Added crash safe tests to make sure slave_gtid_info is updated
correctly even in cases of crashes at specific debug points. Added crash
recovery tests both in MTS and single threaded replication and tested sql_thread
doesn't fail with duplicate key error and missed gtids are printed in the binary
log. Tested by repeatedly crashing a slave with new crash safe parameters
(need to add a rpm test).

Reviewers: jtolmer, rudradevbasak

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D852204

Task ID: 2414895