Fix CREATE TABLE ... LIKE to handle KEY_BLOCK_SIZE
Summary:
When doing a CREATE TABLE ... LIKE where the origin table uses both
ROW_FORMAT=COMPRESSED and a KEY_BLOCK_SIZE which is not the default,
then the created table claims that it is using the same key block as
the origin but in fact uses the default.

SHOW CREATE TABLE will report an incorrect key block size but
innochecksum will show that the file on disk is actually using
the default.

I entered this as http://bugs.mysql.com/bug.php?id=66738.

I also corrected the output of innochecksum so that it doesn't
report things like 4096K.

Internal:
This is a port of https://phabricator.fb.com/D569245.

There are slight changes to the test from 5.1:
  1) innochecksum doesn't support the -b switch in 5.6.
  2) The output for compressed page size is 4096 instead of 4k.

Test Plan: Wrote a new test

Reviewers: rongrong

Reviewed By: rongrong

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D809871

Task ID: 1817762

Conflicts:
	mysql-test/mysql-test-run.pl