Port v5.1 "Convert UNIV_ZIP_DEBUG into sysvar innodb_zip_debug" to 5.6
Summary:
When UNIV_ZIP_DEBUG is enabled, compressed pages are verified after
operations that modify the compressed page. This is done to make sure that decompressing
a compressed page always gives the actual uncompressed page. This functionality was
enabled only if mysql was built using UNIV_ZIP_DEBUG flag. In this diff, we get rid
of this compile time variable and make it a system variable called
innodb_zip_debug that can be enabled at any point.

The motivations for this change are:
1- It would be nice to have Jenkins run the tests with validation turned
on especially while we are changing the compression format.
2- Sometimes a server crashes and it is left intact for inspection.
Often the reason for crash is because of some discrepancy between the compressed and the
uncompressed page. In order to verify whether this is the case, one would need to build a
mysqld with UNIV_ZIP_DEBUG enabled. This is a painful procedure and we have many
instances of xtrabackup failures where we may just enable innodb_zip_debug to see if
the problem has to do with the inconsistency between the compressed and the
uncompressed page.

I don't think this diff would cause too much performance regression if
any because for all of the call sites to page_zip_validate(), I made sure that the call is inside
an if-clause which we give a hint to compiler to discard during branch prediction.

I created innodb-zip-debug versions of tests for all the stress tests
except for the hugestress tests. If Jenkins complains about the fact that
these tests take too much time we can either increase the timeouts for
these tests or we can reduce the amount of queries performed by these tests.

If innodb_zip_debug is enabled for a non-debug build of mysql, a warning
is sent to the client.

Internal: This is a direct port of D602928 except minor code-formatting/indentation changes.

Test Plan: Jenkins (arc unit)

Reviewers: rongrong, nizamordulu

Reviewed By: nizamordulu

CC: nizamordulu, jtolmer, mysql-eng@, steaphan

Differential Revision: https://phabricator.fb.com/D839611

Task ID: 2477927