Port v5.1 "Compress until the end of the last record in page_zip_compress_sec()" to 5.6.
Summary:
page_zip_compress_sec() compressed only until the starting
position of the data part of the last record on a secondary index page.
It relied on page_zip_compress() to add the last record's data to the
compression stream while page_zip_compress() was adding the trailing
garbage.

The removal of compression of trailing garbage will be done in
separate diffs, and this diff forms a base for those. There is no
backward or forward data incompatibility because the amount of trailing
garbage compressed doesn't matter. Even if it did, this diff still
compresses all of the garbage it just passes until the end of the last
record before the trailing garbage.

Internal:

This is a direct code port of D598146.

I just modified the summary a bit.

Test Plan: Jenkins (arc unit)

Reviewers: rongrong

Reviewed By: rongrong

CC: jtolmer, nizamordulu, mysql-eng@

Differential Revision: https://phabricator.fb.com/D840741

Task ID: 2479887