Port v5.1 New Compression "pass page_zip object to page_zip_compress_*" to 5.6
Summary:
page_zip_compress_* functions take a lot of parameters because
they need to write to a temporary buffer and they need pointers to the
trailer of the buffer. Instead of passing these pointers we can pass a
temporary page_zip object that points to the temporary compression
buffer and these functions can recompute the trailer pointers
themselves. This adds some overhead but is well worth the cost because
we are going to use the page_zip object for prefix encoding of trx id
and rollback pointers in future diffs.

Internal:
This is a direct port of D591245, except I just made the new code look
similar to the code around it.

Test Plan: Jenkins (arc unit)

Reviewers: rongrong

Reviewed By: rongrong

CC: jtolmer, nizamordulu, mysql-eng@

Differential Revision: https://phabricator.fb.com/D842366

Task ID: 2484463