Port v5.1 New Compression "Refactor page_zip_compress_clust() to handle compression and storage of uncompressed fields separately" to 5.6
Summary:
Title says it all. By encapsulating the storing of uncompressed
fields in a different function we can eliminate the dependancy of prefix
encoding of transaction id and rollback pointers on the compression of
the page. This revision also refactors page_zip_compress_clust() so that
it calls page_zip_compress_clust_rec() to add a new record with no blob
pointers to the compression stream so page_zip_compress_clust() becomes
a wrapper around two smaller functions that handle the compression of
one record.

Internal:

This is a direct port of D591419, except I just made the added code look
similar to the code around, and add information about function input
parameters.

Test Plan: Jenkins (arc unit)

Reviewers: rongrong, jtolmer, nizamordulu

Reviewed By: rongrong

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D842513

Task ID: 2485287