Make InnoDB 5.6 checksums fully backward compatible
Summary:
There are 4 different kinds of checksums in use.
1) crc32c = ut_crc32(..) ^ ut_crc32(..)
2) crc32cfb = ut_crc32(..) + ut_crc32(..)
3) new_csum = ut_fold_binary(..) + ut_fold_binary(..)
4) old_csum = ut_fold_binary(..)

For checking corruptness, we try to match the value of
checksum_field1 (aka "new style" = mach_read_from_4(buf +
FIL_PAGE_SPACE_OR_CHKSUM) ) to all of the following:

In InnoDB5.1 / Innochecksum5.1 : crc32cfb, new_csum
In InnoDB5.6 : crc32c, crc32cfb
In Innochecksum5.6 : crc32c, crc32cfb, new_csum

This patch adds new_csum to the check within InnoDB5.6
So that any page which is not corrupted according to 5.1 is also not
corrupted according to 5.6

Test Plan:
Ran XtraBackup against an instance which uses new_csum, and against
which XtraBackup used to crash. With this patch, it works correctly.

Reviewers: steaphan

Reviewed By: steaphan