Revert "Disable extra checksum check on startup (temporary)."
Summary:
This reverts commit 1b49bd976dccd4b8366e9b774f0dca50fce3c94b.
	(https://phabricator.fb.com/D821577)

This is a clean revert.

Rudy found that we actually have unused data with older "new" checksums
that we don't ever use.  Somehow, we have data that was written by a non-fb
version of mysql.

When xtrabackup2 checked it, it had problems because it checksums all.
It sounds like this failure was just another symptom of the same problem.

Rudy added support for accepting these older ("new") checksums also.

So, this reverts my disabling of the extra check, so Yoshi can confirm.

This reverts an internal diff, so it will also not be public.

@internal

Test Plan: Clean revert.  Will build an rpm and ask Yoshi to test.

Reviewers: jtolmer

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D852564

Task ID: 2510037, 2420628
(cherry picked from commit 6cafd5d68fb50eae1a94527b17755dc2519db37a)